Properly initialize the VkImage in drm format
The vulkan spec requires us to initialize the image with some special
parameters if we want to later get the drm format modifier, which we
need to access share it as a dma_buf.

If we wanted to be good programmers, we'd have to do some form of
format negotiation between steamvr and dxvk, but we aren't. We are lazy
programmers who do the format negotiation out of band, and I've
negotiated with me to do DRM_FORMAT_MOD_LINEAR (0 on my machine), which
should be supported on both sides.

I have added an interface to query the actual image format from the
driver, but it'll always be 0 since that's the only option I'm giving
it.